+++ aepp/config.py
@@ -60,6 +60,7 @@ endpoints = {
     "observability": "/data/infrastructure/observability/insights",
     "destinationAuthoring": "/data/core/activation/authoring",
     "destinationInstance" : "/data/core/activation/disflowprovider",
+    "hygiene" : "/data/core/hygiene",
     "streaming": {
         "inlet": "",
         "collection": "https://dcs.adobedc.net"